fixing rest endpoint working.
